Add memory simulation module
There is no point in adding memory management to the base implementation
of Obsidian. Yet it is useful to include some way to simulate memory
management. As a result memory management is included as a separate
module that includes all memory functionality while still being
separated from the rest of the Obsidian codebase.